Split `MatrixRoom` into `JoinedMatrixRoom`
This new interface is the one which has guaranteed access to a live timeline, while the `MatrixRoom` one has access to a subset of features rooms both in joined and not joined can access.
